tapis/benchmark/qa_evaluator/pyproject.toml
[project]
name = "qa-evaulator"
version = "0.1.0"
description = ""
authors = [
    {name = "Anagha Jamthe", email = "ajamthe@tacc.utexas.edu"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "nltk >=3.9.1,<4.0.0",
    "rouge-score >=0.1.2,<1.0.0",
    "code-bert-score >=0.3.4,<1.0.0",
    "pandas >=2.2.0,<3.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"